Bug was fixed
Closes #11, #12
