--- core/templates/core/parts/js.html
@@ -1,4 +1,0 @@
-{% load static %}
-
-<!-- Mobile menu -->
-<script defer src="{% static 'js/button.js' %}"></script>
