+++ terra_aurum/settings.py
@@ -1,8 +1,7 @@
 import os
-from pathlib import Path
 
-BASE_DIR = Path(__file__).resolve().parent.parent
 
+BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 SECRET_KEY = os.getenv('KEY', 'development-key')
 DEBUG = os.getenv('DEBUG', False)
 
@@ -57,7 +56,7 @@ WSGI_APPLICATION = 'terra_aurum.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': BASE_DIR / 'db.sqlite3',
+        'NAME': os.path.join(BASE_DIR, 'db.sqlite3')
     }
 }
